backend/requirements.txt
pip
wheel
django
djangorestframework
